Working version of map updates (more testing needed), added homography filtering
